[gnucash/report/reports/standard] introduce owner-billing report
Background:
In Germany, a co-owners' association (MEG) is free to
decide to self-manage the property in an owner administration.
The tasks of property management can then be taken over by an
owner of the MEG appointed for this purpose (introduction link:
https://www.hausverwaltung-ratgeber.de/hausverwaltung.html).

Implementation:
With the current version of 'gnucash', the property management can
create the account structure required for this and manage the
necessary bookings. The legal regulations that exist in Germany
stipulate that individual accounts must be created annually for the
owners, which in particular must show and proportionately allocate
apportionable amounts, non-apportionable amounts, the maintenance
reserve, as well as income, individual and heating costs.

The aim of this new report is to provide such an individual statement.

Signed-off-by: Ralf Zerres <ralf.zerres@networkx.de>